requirements.dev.txt
# Run unittest
pytest==8.4.1
pytest-django==4.11.1

# Monitoring Celery
flower==2.0.1
